feat: enable -r on PS5 and higher
feat(VimCompletion): remove -T, which is not useful on Windows and
Invoke-Vim cannot test because the session is not a terminal.

feat(VimCompletion): use Register-ArgumentCompleter for PS 5.x and
higher. TabExpansion still does much of the completion in PS 5.x, but
Register-ArgumentCompleter prevents interception by TabExpansion2
default file completion, a problem which appeared as of

Name                           Value
----                           -----
PSVersion                      5.1.19041.1
PSEdition                      Desktop
PSCompatibleVersions           {1.0, 2.0, 3.0, 4.0...}
BuildVersion                   10.0.19041.1
CLRVersion                     4.0.30319.42000
WSManStackVersion              3.0
PSRemotingProtocolVersion      2.3
SerializationVersion           1.1.0.1

feat(Get-VimOption): remove -T, which is not useful on Windows and
Invoke-Vim cannot test because the session is not a terminal.

feat(Get-VimSwapFile): use Invoke-Vim for stderr redirection.

test(TabExpansion2): remove -T, which is not useful on Windows and
Invoke-Vim cannot test because the session is not a terminal.

test(TabExpansion2):remove PS 6 or higher version guard.

test(TabExpansion2): remove SkipPS5.

test(TabExpansion): remove -T, which is not useful on Windows and
Invoke-Vim cannot test because the session is not a terminal.

test(TabExpansion): remove SkipPS5.